feat: home page without top 5
