+++ src/pages/HomePage/Home.module.scss
@@ -1,140 +1,18 @@
-.content {
-  padding: 0;
+.contentWrapper {
+  margin-top: 116px;
 }
 
-.root {
-  padding: 24px 50px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  max-width: 930px;
-  margin: auto auto;
-
-  min-height: calc(100vh - 63px) !important;
-  @media screen and (max-width: 767px) {
-    height: calc(100vh - 153px);
-  }
-}
-
-.logo {
-  width: 100%;
-  user-select: none;
-  -webkit-user-drag: none;
-  margin-bottom: 20px;
-}
-
-.title {
-  font-size: 72px !important;
-  font-weight: 400 !important;
+h1 {
+  margin-bottom: 40px;
   text-align: center;
-  position: relative;
-  margin-bottom: 24px !important;
-
-  @media screen and (max-width: 767px) {
-    font-size: 32px !important;
-  }
+  font-style: normal;
+  font-weight: 500;
+  font-size: 104px;
+  line-height: 100%;
 }
 
-.zeroFees {
-  background: #eed600;
-  padding: 8px 24px;
-  font-weight: 400;
-  font-size: 16px;
-
-  color: var(--black-color);
-}
-
-.solanaWord {
-  font-weight: 700 !important;
-  text-transform: uppercase;
-  background: linear-gradient(45deg, #dc1fff, #00ffa3);
-  background-clip: text;
+.emphasis {
+  background: -webkit-linear-gradient(0.5turn, #00FFA3, #DC1FFF);
   -webkit-background-clip: text;
   -webkit-text-fill-color: transparent;
-}
-
-.btn {
-  font-size: 40px;
-  height: unset;
-  background: #94c5fd;
-  padding: 2px 15px;
-  color: var(--black-color);
-  border: unset;
-  border-radius: 20px;
-  border-color: #94c5fd;
-  border: 1px solid #94c5fd;
-  box-shadow: 0px 2px 0px rgba(0, 0, 0, 0.043);
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  cursor: pointer;
-  transition: background var(--ease-base-out) var(--animation-duration-base),
-    border-color var(--ease-base-out) var(--animation-duration-base);
-  margin-bottom: 18px;
-
-  svg {
-    fill: black;
-  }
-
-  &:hover {
-    background: #d1e4ff;
-    border-color: #d1e4ff;
-  }
-
-  @media screen and (max-width: 767px) {
-    font-size: 24px;
-  }
-}
-
-.stats {
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 16px;
-  margin-bottom: 32px;
-
-  @media screen and (max-width: 767px) {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-  }
-}
-
-.statsBlock {
-  background: #2a3139;
-  border: 1px solid #3e4854;
-  border-radius: 20px;
-  padding: 12px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  gap: 8px;
-  p {
-    margin-bottom: 0;
-  }
-
-  @media screen and (max-width: 767px) {
-    width: calc(50% - 8px);
-  }
-}
-
-.statsLabel {
-  font-weight: 700;
-  font-size: 14px;
-  color: #8e9197;
-  text-align: center;
-}
-
-.statsNumber {
-  font-size: 16px;
-  display: flex;
-  align-items: center;
-  gap: 5px;
-}
-
-.solLogo {
-  width: 12px;
-  height: 12px;
 }
